classifier/requirements.txt
click
flask
numpy
Pillow
